requirements.txt
solana
solders
websockets
aiohttp
rich
python-dotenv
requests
twikit
uvloop; sys_platform != "win32"
